requirements.txt
numpy
scikit-learn
tensorflow_addons
opencv-python
pandas
matplot
tqdm
Pillow
scipy